requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.20.0
redis==5.0.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dotenv==1.0.0
scalar-fastapi==1.5.0
email-validator>=2.0.0
bcrypt==4.2.0  # Versión específica para evitar el error __about__
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
alembic==1.13.2
